backend/requirements.txt
aiokafka
aiohttp
python-socketio
fastapi
uvicorn
